+++ src/simulation/simulation_manager.py
@@ -1,4 +1,4 @@
-from characters.character_builder import CharacterFactory
+from characters.character_factory import CharacterFactory
 from characters.enemy_base import Enemy
 from combat.combat_manager_base import CombatManager
